+++ src/env.rs
@@ -1,9 +1,8 @@
-#![allow(dead_code)]
-
 use std::{collections::HashMap};
 
-use crate::types::Type;
+use crate::types::{Args, Ret, Type};
 
+#[derive(Debug)]
 pub struct Env<'a> {
     data: HashMap<String, Type>,
     outer: Option<&'a Env<'a>>,
@@ -15,6 +14,15 @@ impl<'a> Env<'a> {
             data: HashMap::new(),
             outer: outer,
         }
+    }
+
+    pub fn new_default() -> Env<'a> {
+        let mut env = Env::new(None);
+        env.set("+", Type::Fun(sum));
+        env.set("-", Type::Fun(sub));
+        env.set("*", Type::Fun(mul));
+        env.set("/", Type::Fun(div));
+        env
     }
 
     pub fn set(&mut self, symbol: &str, value: Type) {
@@ -43,5 +51,46 @@ impl<'a> Env<'a> {
             },
             None => Err(format!("Symbol '{}' not found in any environment", symbol)),
         }
+    }
+}
+
+
+fn sum(args: Args) -> Ret {
+    match (args.get(0), args.get(1)) {
+        (Some(Type::Int(a)), Some(Type::Int(b))) => Ok(Type::Int(*a + *b)),
+        (Some(Type::Float(a)), Some(Type::Int(b))) => Ok(Type::Float(*a + *b as f64)),
+        (Some(Type::Int(a)), Some(Type::Float(b))) => Ok(Type::Float(*a as f64 + *b)),
+        (Some(Type::Float(a)), Some(Type::Float(b))) => Ok(Type::Float(*a + *b)),
+        _ => Err(String::from("'+' operation failed. Types must be numeric (Int or Float)")),
+    }
+}
+
+fn sub(args: Args) -> Ret {
+    match (args.get(0), args.get(1)) {
+        (Some(Type::Int(a)), Some(Type::Int(b))) => Ok(Type::Int(*a - *b)),
+        (Some(Type::Float(a)), Some(Type::Int(b))) => Ok(Type::Float(*a - *b as f64)),
+        (Some(Type::Int(a)), Some(Type::Float(b))) => Ok(Type::Float(*a as f64 - *b)),
+        (Some(Type::Float(a)), Some(Type::Float(b))) => Ok(Type::Float(*a - *b)),
+        _ => Err(String::from("'-' operation failed. Types must be numeric (Int or Float)")),
+    }
+}
+
+fn mul(args: Args) -> Ret {
+    match (args.get(0), args.get(1)) {
+        (Some(Type::Int(a)), Some(Type::Int(b))) => Ok(Type::Int(*a * *b)),
+        (Some(Type::Float(a)), Some(Type::Int(b))) => Ok(Type::Float(*a * *b as f64)),
+        (Some(Type::Int(a)), Some(Type::Float(b))) => Ok(Type::Float(*a as f64 * *b)),
+        (Some(Type::Float(a)), Some(Type::Float(b))) => Ok(Type::Float(*a * *b)),
+        _ => Err(String::from("'*' operation failed. Types must be numeric (Int or Float)")),
+    }
+}
+
+fn div(args: Args) -> Ret {
+    match (args.get(0), args.get(1)) {
+        (Some(Type::Int(a)), Some(Type::Int(b))) => Ok(Type::Int(*a / *b)),
+        (Some(Type::Float(a)), Some(Type::Int(b))) => Ok(Type::Float(*a / *b as f64)),
+        (Some(Type::Int(a)), Some(Type::Float(b))) => Ok(Type::Float(*a as f64 / *b)),
+        (Some(Type::Float(a)), Some(Type::Float(b))) => Ok(Type::Float(*a / *b)),
+        _ => Err(String::from("'/' operation failed. Types must be numeric (Int or Float)")),
     }
 }