+++ changelog.rst
@@ -5,6 +5,7 @@ CHANGELOG
 -----
 - Add env module with `Env` type
 - Use env as REPL environment instead of a HashMap
+- Pass REPL env as argument to `rep` function
 
 0.4.0
 -----